+++ vulkan/src/VulkanObj/VulkanObj.hpp
@@ -53,7 +53,7 @@ class VulkanObj
 		uint32_t			m_FramesBuffered = 1;
 		uint32_t			m_FrameNumber = 0;
 		bool				m_FramebufferResized = false;
-		uint32_t			m_Numparticles=0;
+		uint64_t			m_Numparticles=0;
 		bool				m_DeviceTimersOn = true;
 		uint32_t			m_Collisons=0;
 		uint32_t			m_SideLength = 0;
